pyproject.toml
[tool.poetry]
name = "devteam"
version = "0.1.0"
description = "Claude-based multi-agent development system"
authors = ["DevTeam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.39.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
httpx = "^0.28.0"
python-telegram-bot = "^21.8"
aiohttp = "^3.9.0"
pygithub = "^2.5.0"
asyncpg = "^0.30.0"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
click = "^8.1.7"
aiofiles = "^24.1.0"
jinja2 = "^3.1.4"
pyyaml = "^6.0.2"
psutil = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true